UserAuthentication/UserAuthentication/UserAuthenticator.cs
using NUnit.Framework;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace UserAuthentication
{
   public class UserAuthenticator
    {
        public bool RegisterUser(string username, string password)
        {
            // Implementation code for user registration
            // ...
            return true; // Return true if registration is successful, false otherwise
        }

        public bool LoginUser(string username, string password)
        {
            // Implementation code for user login
            // ...
            return true; // Return true if login is successful, false otherwise
        }

        public bool ResetPassword(string username, string newPassword)
        {
            // Implementation code for password reset
            // ...
            return true; // Return true if password reset is successful, false otherwise
        }
    }
}

